Refactor factory methods
Extract final fields to allow reuse across iterations

+++ src/main/java/org/sharetrace/experiment/BaseFileExperiment.java
@@ -31,7 +31,7 @@ abstract class BaseFileExperiment implements Experiment {
 
   @Value.Parameter
   @Value.Default
-  protected int numIterations() {
+  protected short numIterations() {
     return 1;
   }
 }
